06 Authentication with GraphQL | 16 Locking Down Mutations (Posts and Comments)

+++ 02-graphql-prisma/src/resolvers/Mutation.js
@@ -95,28 +95,42 @@ const Mutation = {
 
     if (!postExists) { throw new Error('Unable to delete post'); }
 
-
     return prisma.mutation.deletePost({
       where: {
         id: args.id
       }
     }, info);
   },
-  async updatePost(parent, args, { prisma }, info){
+  async updatePost(parent, args, { prisma, request }, info){
+
+    const userId = getUserId(request);
+
+    const postExists = await prisma.exists.Post({
+      id: args.id,
+      author: {
+        id: userId
+      }
+    });
+
+    if (!postExists) { throw new Error('Unable to update post'); }
+
     return prisma.mutation.updatePost({
       where: {
-        id: args.id
+        id: userId
       },
       data: args.data
     }, info);
   },
-  async createComment(parent, args, { prisma }, info) {
+  async createComment(parent, args, { prisma, request }, info) {
+
+    const userId = getUserId(request);
+
     return prisma.mutation.createComment({
       data: {
         text: args.data.text,
         author: {
           connect: {
-            id: args.data.author
+            id: userId
           }
         },
         post: {
@@ -127,14 +141,38 @@ const Mutation = {
       }
     }, info);
   },
-  async deleteComment(parent, args, { prisma }, info) {
+  async deleteComment(parent, args, { prisma, request }, info) {
+
+    const userId = getUserId(request);
+
+    const commentExists = await prisma.exists.Comment({
+      id: args.id,
+      author: {
+        id: userId
+      }
+    });
+
+    if (!commentExists) { throw new Error('Unable to delete comment'); }
+
     return prisma.mutation.deleteComment({
       where: {
         id: args.id
       }
     }, info);
   },
-  async updateComment(parent, args, { prisma }, info) {
+  async updateComment(parent, args, { prisma, request }, info) {
+
+    const userId = getUserId(request);
+
+    const commentExists = await prisma.exists.Comment({
+      id: args.id,
+      author: {
+        id: userId
+      }
+    });
+
+    if (!commentExists) { throw new Error('Unable to update comment'); }
+
     return prisma.mutation.updateComment({
       where: {
         id: args.id
